Rename file to match component
