+++ src/components/navigation/nav-menu/components/nav-icon-item/index.ts
@@ -1,2 +1,2 @@
-export type { NavigationLogoProps } from './nav-icon-item.types';
+export type { NavigationLogoProps } from './navigation-logo.types';
 export { NavigationLogo } from './navigation-logo';
